@@ -1,24 +1,32 @@
 buildscript {
     repositories {
         jcenter()
+        maven {
+            url 'https://maven.google.com/'
+            name 'Google'
+        }
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:1.5.0'
+        classpath 'com.android.tools.build:gradle:3.2.0'
     }
 }
 
 allprojects {
     repositories {
         jcenter()
+        maven {
+            url 'https://maven.google.com/'
+            name 'Google'
+        }
     }
 }
 
 ext {
     minSdkVersion = 8
     sampleMinSdkVersion = 15
-    targetSdkVersion = 23
-    compileSdkVersion = 23
-    buildToolsVersion = '23.0.1'
+    targetSdkVersion = 28
+    compileSdkVersion = 28
+    buildToolsVersion = '28.0.2'
 }
 
 ext.deps = [
@@ -29,6 +37,6 @@ ext.deps = [
         cardview             : 'com.android.support:cardview-v7:23.1.0',
 
         // External
-        recyclerviewanimators: 'jp.wasabeef:recyclerview-animators:1.3.0',
+        recyclerviewanimators: 'jp.wasabeef:recyclerview-animators:2.3.0',
         smoothprogressbar    : 'com.github.castorflex.smoothprogressbar:library-circular:1.1.0'
 ]
