@@ -1,6 +1,6 @@
 package com.paginate;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.widget.AbsListView;
 
 import com.paginate.abslistview.AbsListViewPaginate;
