@@ -1,7 +1,7 @@
 package com.paginate.recycler;
 
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 class DefaultLoadingListItemSpanLookup implements LoadingListItemSpanLookup {
 
