@@ -1,6 +1,6 @@
 package com.paginate.recycler;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -11,7 +11,7 @@ import com.paginate.R;
 public interface LoadingListItemCreator {
 
     /**
-     * Create new loading list item {@link android.support.v7.widget.RecyclerView.ViewHolder}.
+     * Create new loading list item {@link RecyclerView.ViewHolder}.
      *
      * @param parent   parent ViewGroup.
      * @param viewType type of the loading list item.
